Add GameController
Made the gamecontroller to hold the map generator which allows for
multiple maps to be made (for when I want to add multiplayer).
The game controller allows the player to ask what map they're on.
  This needs to allow for cases where the player is not on a map.

Decoupled the TileMarker from the map generation.
TileMarker is now a child of the player.